requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.32.1
jinja2==3.1.6
python-multipart==0.0.19
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
chromadb==1.0.12
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
aiofiles==23.2.1
numpy<2.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.4.0
langchain==0.3.7
langchain-text-splitters==0.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0
pytest-xdist==3.5.0
pytest-env==1.1.3
fastapi-csrf-protect~=1.0.3
openai~=1.86.0